requirements.txt

PyYAML
httpx
loguru
tabulate

pytest
pytest_httpx
